+++ CHANGES.rst
@@ -1,4 +1,8 @@
+0.9.3
+-----
+- Added support for datetime attributes
+- Replaced feature.SetField with feature.SetField2
+
 0.9.2
 -----
-
 - Ported code from geometry_store